models and controllers added

+++ MultiShop.Order.Domain/Entities/OrderDetail.cs
@@ -14,7 +14,7 @@ namespace MultiShop.Order.Domain.Entities
         public decimal ProductPrice { get; set; }
         public decimal ProductAmount { get; set; }
         public decimal ProductTotalPrice { get; set; }
-        public decimal OrderingId { get; set; }
+        public int OrderingId { get; set; }
         public Ordering Ordering { get; set; }
     }
 }
