+++ MultiShop.Order.WebApi/Features/CQRS/Commands/OrderDetailCommands/CreateOrderDetailCommand.cs
@@ -13,6 +13,6 @@ namespace MultiShop.Order.Application.Features.CQRS.Commands.OrderDetailCommands
         public decimal ProductPrice { get; set; }
         public decimal ProductAmount { get; set; }
         public decimal ProductTotalPrice { get; set; }
-        public decimal OrderingId { get; set; }
+        public int OrderingId { get; set; }
     }
 }
